Defensive coding on old Coin Index
Implements #772319848111004 and #772319848111006
